+++ packages/components-js/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `Textarea` component
+
 ## [1.0.2] (2020-02-04)
 
 ### Fixed